requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
uritemplate
inflection
PyYAML